requirements.txt
requests
beautifulsoup4
pdfplumber
python-dotenv
loguru
telebot
apscheduler
